apply code style rules to welcome screen and split components to separate files
